fix: (core) stackblitz files for dropdown (#1596)
* fixed stackblitz files

* fix angular.json file
